+++ src/hershey_cairo.cpp
@@ -1,6 +1,6 @@
 #include "hershey_cairo.hpp"
 
-void draw_glyph(cairo_t *cr, Hershey::Glyph glyph, int x0, int y0, int scale) {
+void draw_glyph(cairo_t *cr, Hershey::Glyph glyph, float x0, float y0, float scale) {
 
     int x1, x2, y1, y2;
     for(int v = 0; v < glyph.get_nvert() - 1; ++v) {
@@ -17,7 +17,7 @@ void draw_glyph(cairo_t *cr, Hershey::Glyph glyph, int x0, int y0, int scale) {
     }
 }
 
-void draw_glyphs(cairo_t *cr, Hershey::String string, int x0, int y0, int scale) {
+void draw_glyphs(cairo_t *cr, Hershey::String string, float x0, float y0, float scale) {
 
     for(int c = 0; c < string.get_nglyph(); ++c) {
         draw_glyph(cr, string.get_glyph(c), x0, y0, scale);
